+++ library/src/main/kotlin/kotlinx/serialization/csv/decode/CollectionCsvDecoder.kt
@@ -3,7 +3,7 @@ package kotlinx.serialization.csv.decode
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.csv.Csv
 import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.CompositeDecoder
+import kotlinx.serialization.encoding.CompositeDecoder.Companion.DECODE_DONE
 
 /**
  * Decode collections (e.g. lists, sets, maps).
@@ -22,7 +22,7 @@ internal class CollectionCsvDecoder(
     override fun decodeSequentially(): Boolean = true
 
     override fun decodeElementIndex(descriptor: SerialDescriptor): Int = when {
-        reader.isDone || elementIndex >= descriptor.elementsCount -> CompositeDecoder.DECODE_DONE
+        reader.isDone || elementIndex >= descriptor.elementsCount -> DECODE_DONE
         else -> elementIndex
     }