+++ library/src/test/kotlin/kotlinx/serialization/csv/config/CsvDelimiterTest.kt
@@ -2,14 +2,13 @@ package kotlinx.serialization.csv.config
 
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.csv.Csv
-import kotlinx.serialization.csv.CsvConfiguration
 import kotlinx.serialization.csv.records.ComplexRecord
 import kotlinx.serialization.csv.records.Enum
 import kotlinx.serialization.test.assertEncodeAndDecode
 import kotlin.test.Test
 
 /**
- * Test [Csv] with different [CsvConfiguration.delimiter]s.
+ * Test [Csv] with different [CsvConfig.delimiter]s.
  */
 @OptIn(ExperimentalSerializationApi::class)
 class CsvDelimiterTest {
@@ -34,11 +33,9 @@ class CsvDelimiterTest {
     )
 
     @Test
-    fun testComma() = Csv(
-        CsvConfiguration(
-            delimiter = ','
-        )
-    ).assertEncodeAndDecode(
+    fun testComma() = Csv {
+        delimiter = ','
+    }.assertEncodeAndDecode(
         "-150,-1,42,9223372036854775807,-2.0,24.24,true,testing,,kotlin.Unit,FIRST",
         ComplexRecord(
             -150,
@@ -57,11 +54,9 @@ class CsvDelimiterTest {
     )
 
     @Test
-    fun testColon() = Csv(
-        CsvConfiguration(
-            delimiter = ';'
-        )
-    ).assertEncodeAndDecode(
+    fun testColon() = Csv {
+        delimiter = ';'
+    }.assertEncodeAndDecode(
         "-150;-1;42;9223372036854775807;-2.0;24.24;true;testing;;kotlin.Unit;FIRST",
         ComplexRecord(
             -150,
@@ -80,11 +75,9 @@ class CsvDelimiterTest {
     )
 
     @Test
-    fun testDot() = Csv(
-        CsvConfiguration(
-            delimiter = '.'
-        )
-    ).assertEncodeAndDecode(
+    fun testDot() = Csv {
+        delimiter = '.'
+    }.assertEncodeAndDecode(
         "-150.-1.42.9223372036854775807.\"-2.0\".\"24.24\".true.testing..\"kotlin.Unit\".FIRST",
         ComplexRecord(
             -150,
@@ -103,11 +96,9 @@ class CsvDelimiterTest {
     )
 
     @Test
-    fun testTab() = Csv(
-        CsvConfiguration(
-            delimiter = '\t'
-        )
-    ).assertEncodeAndDecode(
+    fun testTab() = Csv {
+        delimiter = '\t'
+    }.assertEncodeAndDecode(
         "-150\t-1\t42\t9223372036854775807\t-2.0\t24.24\ttrue\ttesting\t\tkotlin.Unit\tFIRST",
         ComplexRecord(
             -150,